Move code related to client to its own module
